object_detection_mcp/pyproject.toml
[project]
name = "object-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "minio>=7.2.15",
    "pillow==10.4.0",
    "ultralytics==8.3.65",
]
